Merge branch 'master' of https://github.com/SupSuper/OpenXcom into oXc_kL

Conflicts:
	Makefile.am
	src/Battlescape/BattlescapeState.cpp
	src/Battlescape/InventoryState.cpp
	src/Battlescape/MedikitState.cpp
	src/Battlescape/MiniMapState.cpp
	src/Battlescape/ScannerState.cpp
	src/Battlescape/UnitInfoState.cpp
	src/CMakeLists.txt
	src/Engine/Font.cpp
	src/Engine/Font.h
	src/Geoscape/BuildNewBaseState.cpp
	src/Geoscape/Globe.cpp
	src/Geoscape/Globe.h
	src/Interface/Text.cpp
	src/Interface/Text.h
	src/Menu/StartState.cpp
	src/Menu/StartState.h
	src/OpenXcom.2010.vcxproj
	src/Resource/XcomResourcePack.cpp
	src/main.cpp

+++ src/Engine/AdlibMusic.cpp
@@ -114,7 +114,7 @@ void AdlibMusic::load(const std::string& filename)
 	}
 
 	file.seekg(0, std::ifstream::end);
-	_size = file.tellg();
+	_size = static_cast<size_t>(file.tellg());
 	file.seekg(0);
 
 	_data = new char[_size];
